+++ pkg/deployments/ci/prepare-config.ts
@@ -31,6 +31,9 @@ if (process.env.CI) {
       "goerli": {
         "url": "${process.env.GOERLI_RPC_ENDPOINT}"
       },
+      "sepolia": {
+        "url": "${process.env.SEPOLIA_RPC_ENDPOINT}"
+      },
       "zkevm": {
         "url": "${process.env.ZKEVM_RPC_ENDPOINT}"
       }