+++ pkg/deployments/src/types.ts
@@ -5,7 +5,18 @@ export { Artifact, Libraries } from 'hardhat/types';
 
 import Task from './task';
 
-export const NETWORKS = ['goerli', 'mainnet', 'polygon', 'arbitrum', 'optimism', 'gnosis', 'bsc', 'avalanche', 'zkevm'];
+export const NETWORKS = [
+  'goerli',
+  'mainnet',
+  'polygon',
+  'arbitrum',
+  'optimism',
+  'gnosis',
+  'bsc',
+  'avalanche',
+  'sepolia',
+  'zkevm',
+];
 
 export type Network = (typeof NETWORKS)[number];