+++ pkg/deployments/tasks/20220517-protocol-fee-withdrawer/input.ts
@@ -38,6 +38,9 @@ export default {
   goerli: {
     InitialDeniedTokens: [],
   },
+  sepolia: {
+    InitialDeniedTokens: [],
+  },
   zkevm: {
     InitialDeniedTokens: [],
   },