+++ pkg/deployments/tasks/20221123-pool-recovery-helper/input.ts
@@ -57,6 +57,9 @@ export default {
   avalanche: {
     InitialFactories: [],
   },
+  sepolia: {
+    InitialFactories: [],
+  },
   zkevm: {
     InitialFactories: [],
   },